worker_agent/requirements.txt
attrdict
fastapi
httpx
numpy
pillow
pytest
pytest-httpx
python-multipart
tritonclient[all]
uvicorn